+++ src/routes/search.ts
@@ -552,9 +552,14 @@ app.post("/", async (c) => {
   let collection = "offers";
 
   if (
-    ["priceAsc", "priceDesc", "price", "discount", "discountPercent"].includes(
-      sort,
-    )
+    [
+      "priceAsc",
+      "priceDesc",
+      "price",
+      "discount",
+      "discountPercent",
+      "priceUpdatedAt",
+    ].includes(sort)
   ) {
     let priceSortOrder: 1 | -1 =
       sort === "priceAsc" || sort === "priceDesc"
@@ -570,6 +575,10 @@ app.post("/", async (c) => {
 
       if (sort === "discount") {
         return "price.discount";
+      }
+
+      if (sort === "priceUpdatedAt") {
+        return "updatedAt";
       }
 
       return "price.discountPrice";